+++ docs/.vitepress/config.js
@@ -91,6 +91,8 @@ export default defineConfig({
     contributors: {
       merge: 'name',
       debotify: true,
+      resolveGitHub: 'auto',
+      cachePath: 'docs/.vitepress/cache/team-github.json',
       include: [
         {
           name: 'Mike Pirog',